+++ lib/core.scm
@@ -589,4 +589,39 @@
         (set! var1 init1)
         ...
         (let () body1 body2 ...)))))
+(define-syntax syntax-error
+  (er-macro-transformer
+   (lambda (expr rename compare)
+     (apply error (cdr expr)))))
+;;
+;; SRFI-0 (cond-expand) from r7rs
+;;
+
+; TODO: this is just temporary, need to set this in Haskell code (core.hs?)
+;   in order to incorporate system-level fields such as POSIX, windows, etc
+(define *features* '(r7rs husk ratios))
+(define-syntax cond-expand
+  (er-macro-transformer
+   (lambda (expr rename compare)
+     (define (identifier->symbol i) i)
+     (define (check x)
+       (if (pair? x)
+           (case (car x)
+; TODO:              ((and) (every check (cdr x)))
+; TODO:              ((or) (any check (cdr x)))
+             ((not) (not (check (cadr x))))
+; TODO: almost works, but errors out if library not found
+; EG: (cond-expand ((library (base2)) (write "test")) (else (write "none found")))
+             ((library) (eval `(find-module ',(cadr x)) *meta-env*))
+             (else (error "cond-expand: bad feature" x)))
+           (memq (identifier->symbol x) *features*)))
+     (let expand ((ls (cdr expr)))
+       (cond ((null? ls))  ; (error "cond-expand: no expansions" expr)
+             ((not (pair? (car ls))) (error "cond-expand: bad clause" (car ls)))
+             ((eq? 'else (identifier->symbol (caar ls)))
+              (if (pair? (cdr ls))
+                  (error "cond-expand: else in non-final position")
+                  `(,(rename 'begin) ,@(cdar ls))))
+             ((check (caar ls)) `(,(rename 'begin) ,@(cdar ls)))
+             (else (expand (cdr ls))))))))
 ;; END